+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **New**: Easter egg: custom emoji added
 - **New**: Copy only link added to share modal
 - **Fix**: The winning word cannot be a proper Roman numeral (as they often end up in dictionaries)
+- **New**: Keyboard layout picker added to the About language
 
 #### Version 3.18.1 (12.03.2024)
 - **Fix**: Align keyboard heatmaps when the title breaks the line